third_party/blink/web_tests/fast/canvas/canvas-createImageBitmap-data-preserves-alpha-expected.txt

Tests that non-opaque pixels are faithfully preserved in conversions between ImageData and ImageBitmap

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS newImageData.data[0] is within 5 of imageData.data[0]
PASS newImageData.data[1] is within 5 of imageData.data[1]
PASS newImageData.data[2] is within 5 of imageData.data[2]
PASS newImageData.data[3] is imageData.data[3]
PASS successfullyParsed is true

TEST COMPLETE
